Fix intermittent seg fault and status deletion, add view options, and add filename in status
Fixed a segmentation fault that will crash the program when either the rows or columns is 0, but not both. There was also a bug where the Operation options would not work properly on Status.

Menu options were added to help view each table.

A status for the current filename is added to help the user figure out what file they have open.